requirements.txt
faster-whisper
google-cloud-speech